Added cross-sampling for groups of variables
Added method for cross-sampling based on groups of variables.

An additional parameter (groups=False) is an input to the function to
specify whether the cross-sampling should be based on groups. The
initial comments in the function were updated.

The number of rows in the cross-sampled matrix depends on whether
groups were used. If there are groups, then the effective number of
parameters is Dg (number of groups), not the number of variables D.

The cross-sampling portion has an if/else statement to accept for if
cross-sampling is based on groups. If it is not based on groups, the
previous calculations are used. If groups are used, then only the
“diagonal” elements of the matrix (e.g. group 2 in row 2 of the
cross-sampled A(B) matrix).