Added condition for group indices
Added conditions to calculate group indices, which use the same method
of the individual indices but the dimensions of the cross-sampled
matrix must be changed based on the number of groups.

Added the variable ‘groups = False’ to the input arguments, so default
value is not to use groups.

Changed value of ‘D’ to be the number of variables if groups = False or
the number of groups if groups = True.