Added conversion to non-uniform distributions
Added code to convert the 0-to-1 variables from the saltelli function
into other distributions by using the inverse CDF.

A parameter ‘dists’ was added to the function, so when ‘dists=None’ the
uniform scaling is used (previous code), but if a list of distributions
is passed to the function the variables are converted based on the two
or three parameters listed in the ‘bounds’ entry of the problem.

The possible distributions include uniform, triangular (with or without
a specified most likely value), normal, and log-normal (with or without
a real-space lower bound). The code prints error messages when
distribution parameters are improperly specified (e.g. a normal
distribution with a negative standard deviation).