Added variables to call conversion to non-uniform distributions
Added code to call the ‘scaled_samples’ function properly when
non-uniform distributions are needed.

The parameter ‘dists’ was added to the function call, and if
‘dists=True’ the non-uniform distributions specified in the problem
dictionary will be used. An if/else condition was added towards the end
of ‘sample’ to account for the call to ‘scaled_samples’ if there is or
is not a need for non-uniform distributions.