Changed function for non-uniform scaling and reading-in problem file
Revised code to make conversion to non-uniform distributions a separate
function. This restores the initial ‘scale_samples’ function and adds a
new function, ‘nonuniform_scale_samples’. The change was made so that
functions that called ‘scale_samples’ would not need to account for any
changes to that function.

The function ‘read_param_file’ was modified so that it will accept a
fifth column for the distribution name. The new input files must have
‘NA’ listed for groups when no grouping is required but there are
distributions specified. The function adds a ‘dists’ variable to the
problem dictionary, and when no distributions are specified the value
is None.