Changed to use problem dictionary
The function was restructured so that it will use the problem
dictionary directly for group sampling.

The main changes are in using the matrix part of the ‘groups’ variable
when cross-sampling the matrices.  Also, groups is not passed to the
function directly, but if the entry in the dictionary for ‘groups’ is
None then standard sampling is used.

A second change is the method of converting the variables to other
distributions. This was done to account for the new function added for
non-uniform variables.