Replaced '== None' conditions, allowed problem definition from code
Replaced conditions that were ‘== None’ to be ‘no problem.get()’.

Added a condition where the user can specify the problem dictionary
without reading in a parameter file. The new problem dictionary would
have an entry for ‘groups’, but it would not have the tuple generated
in ‘read_param_file’. The new code recognizes this definition and
creates the tuple.