Remove custom exceptions
I hold the opinion that custom exceptions rarely add value and built-ins are far more accessible and appropriate enough.

You can search online for the debate but here's one thread.

https://www.reddit.com/r/Python/comments/f8xobl/best_practices_for_custom_exceptions/

For the two I removed, I asked myself what would the user gain from these custom exception types that they wouldn't get from the message alone?

If the user is a programmer/modeller seeking to integrate SALib into their work, is it worth the added complexity of them having to understand what these custom exceptions are and how to handle them?

For us as maintainers, is the added complexity and need to flesh out the documentation and specify their appropriate use cases worth the effort when "ValueError" covers this case generally?

I judged the answer to be "no" for two of these - unsure of the third but happy to leave it as is for now.

Also, to explain some other changes: on error messages, I find text in all caps to be something anachronistic. I also avoid directing "blame" (e.g., user error).